Refactor the connection pool to be a singleton
The config class will initialize the connection pool and all other packages reference from there.
This saves the passing around of the pool in every function that needed to interact with the database and improves the readability.
An IndexCache class is used to query the INDEX_CACHE table to move towards single responsibility classes.