Return only files from _get_file_sources
YaraProcessor._get_file_sources checked whether the
current hit was a file, in which case it was returned. If it was
a directory, it returned all hits within that. This was a bug since
the inner hit could have been a directory itself.
This commit checks that the inner hit is indeed a file before yielding
it

Closes: #1